requirements.txt
streamlit
scikit-learn
matplotlib
scikit-image
numpy
Pillow
